+++ metadata.rb
@@ -1,8 +1,8 @@
-name             'ephemeral_lvm-nativex'
-maintainer       'NativeX'
-maintainer_email 'derek.bromenshenkel@nativex.com'
+name             'ephemeral_lvm-simplyadrian'
+maintainer       'simplyadrian'
+maintainer_email 'simplyadrian@gmail.com'
 license          'All rights reserved'
-description      'Wraps ephemeral_lvm for NativeX, where it adds a dependency to xfs cookbook.'
+description      'Wraps ephemeral_lvm for simplyadrian, where it adds a dependency to xfs cookbook.'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'